Updated mmr_attachment_check
Moved from the transition_post_status hook to the save_post one.

Added logic to output the invalid not populated fields on a per image basis

Added two checks not to run the logic when auto saving and if the current user has the rights to run the logic.

Only error messages and admin notices are missing